piglit: Support resuming from a previous run
It can happen that the machine hard hangs in one of tests, say,
drv_module_reload. We really want to support resuming the run in that
case (probably excluding the faulty test as well). This commit adds an
easy way to do that:

sudo make run-tests PIGLIT_FLAGS="-x module_reload" RESUME=20131114-piglit-results.1

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>