piglit: Fix resuming of previous runs
When resuming and loading the partial JSON file from disk, the 'status'
property is replaced by status.Status objects. When serializing back
those objects, JSONEncoder is unhappy because it doesn't know how to
serialize status.Status objects and gives up with exceptions like:

TypeError: skip is not JSON serializable

We can write a small subclass of JSONEncoder that knows about Status
objects and use it to do the initial write back of the partial JSON
file.

This commit has been posted on the piglit ml:

http://lists.freedesktop.org/archives/piglit/2013-November/008447.html

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>