lib: Introduce igt_run_quick()
Turns out that we'd like to be able a "quick" version of the tests on
real hardware as well. So "in simulation" and "quick" are two different
concepts.

Note that "in simulation" implies "quick", so igt_run_quick() tests if
we are running on simulation before looking at the IGT_QUICK env
variable.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>