lib: Build a list of property updates instead of tracking changed states
When changing a property (rotation is the only one on planes atm), let's
build a list of changes instead of individually tracking each property
state. This should bring us a bit closer to the atomic iotcl() (ok, only
by a notch, but that's a start).

This will allow us to add more plane properties we want to change in a
hopefully slightly more future proof framework.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>